pyproject.toml
[tool.black]
line-length = 99

[tool.ruff]
line-length = 99

[tool.flake8]
max-line-length = 99

[tool.ruff.lint]
select = ["E", "F"]

[tool.sqlfluff]
dialect = "postgres"
exclude_rules = "L003"

[project]
name = "osm-caclr-address-dashboard"
version = "0.0.0"
requires-python = ">=3.12"
dependencies = [
    "psycopg2-binary",
    "Jinja2",
    "rich>=14.0.0",
    "pysqlite3-binary>=0.5.4",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "sqlfluff>=3.4.1",
]
